chore: re-write hermione.js to typescript
