+++ tsconfig.spec.json
@@ -1,6 +1,6 @@
 {
     "extends": "./tsconfig.common.json",
-    "include": ["lib", "test", "hermione.js", "gemini.js", "playwright.ts"],
+    "include": ["lib", "test", "hermione.ts", "gemini.js", "playwright.ts"],
     "exclude": ["test/func"],
     "compilerOptions": {
         "baseUrl": ".",
